requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==4.2; python_version >= '3.8'
django-dump-die==0.1.7; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
